+++ tests/Containers/Interfaces/IInventory/IInventoryTests.AddItem.cs
@@ -14,19 +14,5 @@ namespace TheChest.Inventories.Tests.Containers.Interfaces
 
             Assert.That(result, Is.True);
         }
-
-        [Test]
-        public void AddItem_FullInventory_ReturnsFalse()
-        {
-            var size = this.random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
-            var items = this.itemFactory.CreateDefault();
-            var inventory = this.inventoryFactory.FullContainer(size, items);
-
-            var item = this.itemFactory.CreateDefault();
-
-            var result = inventory.Add(item);
-
-            Assert.That(result, Is.False);
-        }
     }
 }